% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect-highways.R
\name{connect_highways}
\alias{connect_highways}
\title{connect_highways}
\usage{
connect_highways(highways, bbox, outer = TRUE)
}
\arguments{
\item{highways}{A vector of highway names passed directly to the Overpass
API. Wildcards and whitespaces are `.'; for other options see online help for
the overpass API.}

\item{bbox}{the bounding box for the map.  A 2-by-2 matrix of 4 elements with
columns of min and max values, and rows of x and y values.}

\item{outer}{Either a single logical parameter, or a logical vector of the
same length as \code{hightways}. For cases where multiple paths are possible, the
default \code{outer = TRUE} will select the path lying farthest to the outside of
the resultant polygon, while \code{outer = FALSE} will select the path
lying closest to the inside.}
}
\value{
A single set of \code{SpatialPoints} containing the lat-lon
coordinates of the cyclic line connecting all given streets.
}
\description{
Takes a list of highways names which must enclose an internal area, and
returns a \code{SpatialLines} object containing a sequence of OSM nodes which
cyclically connect all highways. Will fail if the streets do not form a
cycle.
}
\note{
This function can not be guaranteed failsafe owing both to the
inherently unpredictable nature of OpenStreetMap, as well as to the unknown
relationships between named highways.
}
\examples{
bbox <- get_bbox (c (-0.13, 51.5, -0.11, 51.52))
bbox <- array (bbox, dim = c (2, 2))
dimnames (bbox) <- list (c ("x", "y"), c ("min", "max"))

highways <- c (
    "Monmouth.St", "Short.?s.Gardens", "Endell.St", "Long.Acre",
    "Upper.Saint.Martin"
)
# Note that dots signify "anything", including whitespace and apostrophes,
# and that '?' denotes optional previous character and so here matches
# both "Shorts Gardens" and "Short's Gardens"
\dontrun{
highways1 <- connect_highways (highways = highways, bbox = bbox)
highways <- c ("Endell.St", "High.Holborn", "Drury.Lane", "Long.Acre")
highways2 <- connect_highways (highways = highways, bbox = bbox)
}

}
