% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw-to-polygons.R
\name{hws_to_polygons}
\alias{hws_to_polygons}
\title{Convert all highway lines into polygons bounded by the surrounding polygon
geometry of buildings and open spaces.}
\usage{
hws_to_polygons(osmdat)
}
\arguments{
\item{osmdat}{Object returned from \link{sb_osmdata_extract}.}
}
\value{
An \pkg{sf} \code{data.frame} of polygons tracing polygons around all
interior streets until the edges of surrounding polygons.
}
\description{
Convert all highway lines into polygons bounded by the surrounding polygon
geometry of buildings and open spaces.
}
