% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-inspect.R
\name{no_building_floors}
\alias{no_building_floors}
\title{Identify and optionally browse Open Street Map (OSM) buildings with missing
data on numbers of levels.}
\usage{
no_building_floors(osmdat, browse = TRUE)
}
\arguments{
\item{osmdat}{Main object returned by \link{sb_osmdata_extract} function.}

\item{browse}{If \code{TRUE} (default), open OSM pages for all identified ways in
default web browser.}
}
\description{
Identify and optionally browse Open Street Map (OSM) buildings with missing
data on numbers of levels.
}
