% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-inspect.R
\name{no_parking_ways}
\alias{no_parking_ways}
\title{Identify and optionally browse Open Street Map (OSM) highways with missing
parking data.}
\usage{
no_parking_ways(osmdat, min_len = 20, browse = TRUE)
}
\arguments{
\item{osmdat}{Main object returned by \link{sb_osmdata_extract} function.}

\item{min_len}{Minimal length in metres of ways to inspect. Shorter ways are
likely just connector elements which often neither have not require parking
data.}

\item{browse}{If \code{TRUE} (default), open OSM pages for all identified ways in
default web browser.}
}
\value{
(Invisibly) A \code{data.frame} object with columns of OSM identifier
values for each identified way, and corresponding URLs.
}
\description{
Identify and optionally browse Open Street Map (OSM) highways with missing
parking data.
}
