% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population-estimate.R
\name{sb_car_spaces_per_resident}
\alias{sb_car_spaces_per_resident}
\title{Estimate number of car spaces per resident}
\usage{
sb_car_spaces_per_resident(osmdat)
}
\arguments{
\item{osmdat}{Object returned from \link{sb_osmdata_extract}.}
}
\value{
An estimate of numbers of parking spaces per resident.
}
\description{
Estimate number of car spaces per resident
}
