% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osmdata-extract.R
\name{sb_osmdata_extract}
\alias{sb_osmdata_extract}
\title{Extract all data needed for analyses for an area within a bounding box and
bounded by a series of named highways.}
\usage{
sb_osmdata_extract(bbox, hw_names, outer = TRUE)
}
\arguments{
\item{bbox}{Bounding box as vector of (xmin, ymin, xmax, ymax).A}

\item{hw_names}{Character vector naming highways which entirely surround
desired area. Highway names can be entered in any order.}

\item{outer}{Single value or logical vector of same length as \code{hw_names},
specifying whether perimeter of desired area should be traced along outer
(TRUE) or inner (FALSE) ways.}
}
\value{
A list of data components, each in \pkg{sf} format.
}
\description{
Extract all data needed for analyses for an area within a bounding box and
bounded by a series of named highways.
}
