% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking-km-per-empty.R
\name{sb_parking_km_per_empty}
\alias{sb_parking_km_per_empty}
\title{Total kilometres driven to fill all parking spaces}
\usage{
sb_parking_km_per_empty(osmdat, prop_min = 0.7, n_props = 10, ntrials = 10)
}
\arguments{
\item{osmdat}{Result of \link{sb_osmdata_extract} function, containing all
primary data needed for analyses.}

\item{prop_min}{Minimal proportion of filled parking spaces from which to
start simulation. Low values close to zero and uninformative, as parking
spaces are almost always available directly in front of buildings.}

\item{n_props}{Number of park-filling proportions used to esimtate times.}

\item{ntrials}{Number of random simulations for each value of \code{n_props} used
to generate average values.}
}
\value{
A \code{data.frame} of three columns:
\enumerate{
\item \code{prop_full} The initial proportion of filled parking spaces.
\item \code{d} The resultant average distance per km to fill all remaining spaces
starting at the value specifified by \code{prop_full}.
\item \code{d_to_parking} Equilvalent average distances to drive directly to
nearest large parking facilitity, instead of circling around searching for
spaces within the specified superblock itself.
}
}
\description{
Total kilometres driven to fill all parking spaces
}
