% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking-times.R
\name{sb_parking_times}
\alias{sb_parking_times}
\title{Esimate times to park cars within superblock, compared with times to park in
nearby parking facilities and walk back.}
\usage{
sb_parking_times(osmdat, n_props = 20L, ntrials = 100L, added_time_to_park = 2)
}
\arguments{
\item{osmdat}{Result of \link{sb_osmdata_extract} function, containing all
primary data needed for analyses.}

\item{n_props}{Number of park-filling proportions used to esimtate times.}

\item{ntrials}{Number of random simulations for each value of \code{n_props} used
to generate average values.}
}
\description{
Esimate times to park cars within superblock, compared with times to park in
nearby parking facilities and walk back.
}
