% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superblock-package.R
\docType{package}
\name{superblock-package}
\alias{superblock}
\alias{superblock-package}
\title{superblock: Analyses of potential superblock}
\description{
Analyses of potential superblock
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/superblock-ms}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/superblock-ms/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
