% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hw-to-polygons.R
\name{hws_to_polygons}
\alias{hws_to_polygons}
\title{Convert all highway lines into polygons bounded by the surrounding polygon
geometry of buildings and open spaces.}
\usage{
hws_to_polygons(osmdat, interval = 10)
}
\arguments{
\item{osmdat}{Object returned from \link{sb_osmdata_extract}.}

\item{interval}{in metres to interpolate highways and project onto nearby
polygons. Values of 0 or less will not interpolate, and polygons will only
be defined by using points at which lines are defined within Open Street
Map. This may reduce the accuracy of resultant polygons.}
}
\value{
An \pkg{sf} \code{data.frame} of polygons tracing polygons around all
interior streets until the edges of surrounding polygons.
}
\description{
Convert all highway lines into polygons bounded by the surrounding polygon
geometry of buildings and open spaces.
}
