% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parking-times.R
\name{sb_parking_times}
\alias{sb_parking_times}
\title{Esimate times to park cars within superblock, compared with times to park in
nearby parking facilities and walk back.}
\usage{
sb_parking_times(osmdat, n_props = 10L, prop_min = 0.7, ntrials = 100L)
}
\arguments{
\item{osmdat}{Result of \link{sb_osmdata_extract} function, containing all
primary data needed for analyses.}

\item{n_props}{Number of park-filling proportions used to esimtate times.}

\item{prop_min}{Minimal proportion of filled parking spaces from which to
start simulation. Low values close to zero and uninformative, as parking
spaces are almost always available directly in front of buildings.}

\item{ntrials}{Number of random simulations for each value of \code{n_props} used
to generate average values.}
}
\value{
A \code{data.frame} of proportions of filled parking spaces and
corresponding simulated times to (via car) from (walking) nearest available
spaces. Columns are also included for corresponding times to and from nearby
parking facilities ('to_park' and 'from_park'). all times are in minutes.
}
\description{
Esimate times to park cars within superblock, compared with times to park in
nearby parking facilities and walk back.
}
