% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{sb_summary}
\alias{sb_summary}
\title{Print screen summary of effect of superblock on public space.}
\usage{
sb_summary(osmdat, hw_polygons = NULL, add_parking_osm_ids = NULL)
}
\arguments{
\item{osmdat}{Object returned from \link{sb_osmdata_extract}.}

\item{hw_polygons}{Object returned from \link{hws_to_polygons}.}

\item{add_parking_osm_ids}{Optional list of 'OSM' id values for any ways
which have no parking specified in OpenStreetMap, which are then assigned
average values from all other ways which do have parking specified.}
}
\value{
(Invisibly) A named list of the areas devoted to the aspects
presented in the summary (all in Hectares).
}
\description{
Print screen summary of effect of superblock on public space.
}
